requirements.txt
chromadb==1.0.20
sentence-transformers==5.1.0
pyyaml==6.0.2
streamlit==1.37.1
pandas==2.3.1
plotly==6.3.0
requests==2.32.5
python-docx==1.2.0
reportlab==4.4.3
pypdf==6.0.0
